moves factories folder
