+++ packages/relay/src/lib/factories/transactionFactory.ts
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
-import { nanOrNumberTo0x, numberTo0x, prepend0x, trimPrecedingZeros } from '../../../formatters';
-import constants from '../../constants';
-import { Log, Transaction, Transaction1559, Transaction2930 } from '../../model';
+import { nanOrNumberTo0x, numberTo0x, prepend0x, trimPrecedingZeros } from '../../formatters';
+import constants from '../constants';
+import { Log, Transaction, Transaction1559, Transaction2930 } from '../model';
 
 // TransactionFactory is a factory class that creates a Transaction object based on the type of transaction.
 export class TransactionFactory {
