+++ packages/relay/src/lib/services/ethService/blockService/BlockService.ts
@@ -9,11 +9,11 @@ import { Utils } from '../../../../utils';
 import { MirrorNodeClient } from '../../../clients/mirrorNodeClient';
 import constants from '../../../constants';
 import { predefined } from '../../../errors/JsonRpcError';
+import { BlockFactory } from '../../../factories/blockFactory';
+import { TransactionFactory } from '../../../factories/transactionFactory';
 import { Block, Log, Receipt, Transaction } from '../../../model';
 import { IContractResultsParams, MirrorNodeBlock, RequestDetails } from '../../../types';
 import { CacheService } from '../../cacheService/cacheService';
-import { BlockFactory } from '../../factories/blockFactory';
-import { TransactionFactory } from '../../factories/transactionFactory';
 import { IBlockService, ICommonService } from '../../index';
 import { CommonService } from '../ethCommonService/CommonService';
