+++ packages/relay/src/lib/services/ethService/transactionService/TransactionService.ts
@@ -12,16 +12,16 @@ import { MirrorNodeClient } from '../../../clients/mirrorNodeClient';
 import constants from '../../../constants';
 import { JsonRpcError, predefined } from '../../../errors/JsonRpcError';
 import { SDKClientError } from '../../../errors/SDKClientError';
-import { Log, Transaction } from '../../../model';
-import { Precheck } from '../../../precheck';
-import { ITransactionReceipt, RequestDetails } from '../../../types';
-import { CacheService } from '../../cacheService/cacheService';
-import { TransactionFactory } from '../../factories/transactionFactory';
+import { TransactionFactory } from '../../../factories/transactionFactory';
 import {
   IRegularTransactionReceiptParams,
   ISyntheticTransactionReceiptParams,
   TransactionReceiptFactory,
-} from '../../factories/transactionReceiptFactory';
+} from '../../../factories/transactionReceiptFactory';
+import { Log, Transaction } from '../../../model';
+import { Precheck } from '../../../precheck';
+import { ITransactionReceipt, RequestDetails } from '../../../types';
+import { CacheService } from '../../cacheService/cacheService';
 import HAPIService from '../../hapiService/hapiService';
 import { CommonService, ICommonService } from '../../index';
 import { ITransactionService } from './ITransactionService';
